feat/#44-기존 코드 삭제, 변경된 구조로 반영
